requirements.txt
accelerate
datasets >= 1.8.0
sentencepiece != 0.1.92
transformers
deepspeed
neptune
pdbpp
notebook
protobuf==3.20.*
pyyaml
pynvml
hydra-core
evaluate
nltk
absl-py
rouge_score
